requirements.txt
flask
requests
gunicorn
google-generativeai
PyGithub
